requirements.txt

torch>=2.1
numpy>=1.24
pandas>=2.0
scipy>=1.10
pyyaml>=6.0
psycopg2-binary>=2.9
tqdm>=4.65
networkx>=3.0
